refactor: seperate components and tests from stories
