+++ src/components/SearchBar.test.tsx
