+++ src/components/index.ts
@@ -1,0 +1,4 @@
+export * from "./Button"
+export * from "./Checkbox"
+export * from "./SearchBar"
+export * from "./Table"
